move getDisplayedValue into separate utils file
